Add: All issues are optional.
Added the ability to silence even the issues most fundamental to the tool (imports) as well as some obscure issues that didn't seem worth exposing earlier (naming & meta).  This downgrades a lot of meta issues which previously (always) were considered errors and are now (by default) are considered warnings, in order to bring all of them into alignment to use the same GlobalOption.  This is done so that a team can bring LeanHeaders into their project and can start from literally no checks or can choose to check only e.g. the typedef based issues.

This commit adds backwards-compatible behaviour so the minor version is bumped.